Add spec for promiseWhile
Fixed a bad bug in it's execution. It was designed to reject the promise
when the "worker" function (called `body()`) threw an exception. But
because it was called outside of Q it would not actually resolve.

    Q.when(body(), loop, done.reject);

Can't pass a function to Q.when because it thinks it's a value.

    Q.when(body, loop, done.reject);

Instead what we want is for Q to call the function and then resolve /
reject:

    Q.fcall(body).then(loop, done.reject);