Add inital impress presentation
No content. But used the Yeoman generator to build a presentation. It
uses handlebars to build the steps which are defined in the steps/
folder. An easy way to add new steps is to yse the Yeoman generator like
so:

    $ cd impress
    $ yo impress:step "My New Step Title"

Then edit `steps/list.json` and the new HTML fragment it created.